requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
idna==2.6
praw==5.3.0
prawcore==0.13.0
pytz==2017.3
requests==2.18.4
six==1.11.0
typing==3.6.4
update-checker==0.16
urllib3==1.22
coreapi==2.3.3
coreschema==0.0.4
itypes==1.1.0
uritemplate==3.0.0
django-extensions==1.9.9
django-filter==1.1.0
djangorestframework==3.7.7
